requirements.txt
# =============================================================================
# Med Help USA - Voice Intake Agent
# Requirements File (Tested & Working Versions)
# =============================================================================

# -----------------------------------------------------------------------------
# LiveKit Voice Agent Framework (Core)
# -----------------------------------------------------------------------------
livekit==1.0.18
livekit-agents==1.2.18
livekit-api==1.0.7
livekit-protocol==1.0.8
livekit-plugins-openai==1.2.18
livekit-plugins-silero==1.2.18
livekit-plugins-turn-detector==1.2.18
livekit-plugins-noise-cancellation==0.2.5
livekit-blingfire==1.0.0

# -----------------------------------------------------------------------------
# OpenAI (Realtime Voice API)
# -----------------------------------------------------------------------------
openai==2.7.1

# -----------------------------------------------------------------------------
# Database (Supabase)
# -----------------------------------------------------------------------------
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
postgrest==2.24.0
storage3==2.24.0
realtime==2.24.0

# -----------------------------------------------------------------------------
# Environment & HTTP
# -----------------------------------------------------------------------------
python-dotenv==1.2.1
aiohttp==3.13.2
httpx==0.28.1

# -----------------------------------------------------------------------------
# Audio Processing
# -----------------------------------------------------------------------------
sounddevice==0.5.3
av==15.1.0

# -----------------------------------------------------------------------------
# AI/ML Dependencies
# -----------------------------------------------------------------------------
onnxruntime==1.23.1
transformers==4.57.1
tokenizers==0.22.1
huggingface-hub==0.36.0

# -----------------------------------------------------------------------------
# Core Utilities
# -----------------------------------------------------------------------------
pydantic==2.12.3
numpy
PyYAML==6.0.3
websockets==15.0.1
cryptography==46.0.3

# =============================================================================
# SYSTEM REQUIREMENTS
# =============================================================================
# - Python 3.10 or higher (tested on Python 3.12/3.14)
# - Windows/macOS/Linux supported
# - OpenAI API Key (for Realtime Voice API)
# - Supabase Project (for lead storage)
# - LiveKit Server (local or cloud)
# - Microphone for voice input
#
# =============================================================================
# QUICK START GUIDE
# =============================================================================
#
# STEP 1: Create Virtual Environment (Recommended)
# ------------------------------------
#   python -m venv venv
#   
#   # Windows:
#   .\venv\Scripts\activate
#   
#   # macOS/Linux:
#   source venv/bin/activate
#
# STEP 2: Install Python Dependencies
# ------------------------------------
#   pip install -r requirements.txt
#
# STEP 3: Download LiveKit Server
# ------------------------------------
#   - Go to: https://github.com/livekit/livekit/releases
#   - Download the latest release for your OS:
#     * Windows: livekit_X.X.X_windows_amd64.zip
#     * macOS: livekit_X.X.X_darwin_amd64.tar.gz  
#     * Linux: livekit_X.X.X_linux_amd64.tar.gz
#   - Extract and place executable in project folder
#
# STEP 4: Create .env File
# ------------------------------------
#   Copy .env.example to .env and fill in:
#
#   OPENAI_API_KEY=sk-your-openai-api-key-here
#   SUPABASE_URL=https://your-project.supabase.co
#   SUPABASE_SERVICE_ROLE_KEY=your-service-role-key
#
# STEP 5: Setup Supabase Table (First Time Only)
# ------------------------------------
#   python supabase_setup.py
#   (Follow instructions to create table in Supabase Dashboard)
#
# STEP 6: Start LiveKit Server (Terminal 1)
# ------------------------------------
#   Windows:     .\livekit-server.exe --dev
#   macOS/Linux: ./livekit-server --dev
#
# STEP 7: Run the Voice Agent (Terminal 2)
# ------------------------------------
#   python main.py console
#
# =============================================================================
# TROUBLESHOOTING
# =============================================================================
# 
# Error: "Missing Supabase credentials"
#   -> Create .env file with SUPABASE_URL and SUPABASE_SERVICE_ROLE_KEY
#
# Error: "OpenAI API key not found"
#   -> Add OPENAI_API_KEY to your .env file
#
# Error: "Cannot connect to LiveKit"
#   -> Make sure livekit-server is running with --dev flag
#
# Error: "Module not found"
#   -> Run: pip install -r requirements.txt
#
# Error: "No audio device"
#   -> Connect a microphone and check system audio settings
#
# Error: "PortAudio" or audio issues on Linux
#   -> Run: sudo apt-get install portaudio19-dev python3-pyaudio
#
# =============================================================================
